Refactoring out a base Widget component
